Refactor and add skeleton tests
