+++ babel.config.js
@@ -3,6 +3,7 @@ module.exports = {
     ["@babel/plugin-transform-private-methods", { loose: "false" }],
     ["@babel/plugin-transform-private-property-in-object", { loose: "false" }],
     ["@babel/plugin-transform-class-properties", { loose: "false" }],
+    "babel-plugin-transform-import-meta"
   ],
   "presets": [
     [
