+++ package.json
@@ -67,6 +67,7 @@
     "@testing-library/user-event": "^14.5.2",
     "babel-jest": "^29.6.2",
     "babel-plugin-styled-components": "^2.1.4",
+    "babel-plugin-transform-import-meta": "^2.3.2",
     "babel-preset-gatsby": "^3.14.0",
     "eslint": "^8.57.1",
     "eslint-config-react-app": "^7.0.1",
